Simplify example apps
